v10

+++ classRoom/views.py
@@ -5,6 +5,7 @@ from .forms import NewClassForm, NewContentForm, NewHomeworkForm, _NewHomeworkDe
 from django.core.files.storage import FileSystemStorage
 from .models import NewClass, ClassContent, ClassHomework, Homework, ClassPost
 from django.http import HttpResponseRedirect
+from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 
 import string
@@ -25,6 +26,7 @@ def index(request):
     return render(request, "classes.html")
 
 
+@login_required(login_url="user:login")
 def show(request):
     # cloud = Cloud.objects.all()
     classes = NewClass.objects.filter(class_teacher_name=request.user)
@@ -51,82 +53,97 @@ def addclass(request):
     return render(request, "addclasses.html", {"form": form})
 
 
+@login_required(login_url="user:login")
 def addhomework(request, id):
     classroom = get_object_or_404(NewClass, id=id)
-    # print(classroom.id)
-    class_content = ClassContent.objects.filter(classroom=id)
 
-    _temp = 0
-    temp = id_generator()
+    if classroom.class_teacher_name_id == request.user.id:
+        # print(classroom.id)
+        class_content = ClassContent.objects.filter(classroom=id)
 
-    if request.method == "POST":
+        _temp = 0
+        temp = id_generator()
 
-        for i in class_content:
+        if request.method == "POST":
 
-            form = NewHomeworkForm(request.POST or None)
+            for i in class_content:
 
-            if form.is_valid():
-                homework = form.save(commit=False)
+                form = NewHomeworkForm(request.POST or None)
 
-                homework.class_name = i.class_name
-                homework.student_name = i.student_naame
-                homework.homework_code = temp
-                homework.student_id = i.student_name_id
-                homework.classroom_id = id
+                if form.is_valid():
+                    homework = form.save(commit=False)
 
-                if _temp == 0:
-                    _form = SecondHomework(request.POST or None)
+                    homework.class_name = i.class_name
+                    homework.student_name = i.student_naame
+                    homework.homework_code = temp
+                    homework.student_id = i.student_name_id
+                    homework.classroom_id = id
 
-                    tempHomework = _form.save(commit=False)
+                    if _temp == 0:
+                        _form = SecondHomework(request.POST or None)
 
-                    tempHomework.title = form.cleaned_data.get("title")
-                    tempHomework.content = form.cleaned_data.get("content")
-                    tempHomework.end_date = form.cleaned_data.get("end_date")
-                    tempHomework.end_clock = form.cleaned_data.get("end_clock")
-                    tempHomework.homework_code = temp
-                    tempHomework.classroom_id = id
-                    tempHomework.submit_count = 0
-                    tempHomework.is_end = 0
-                    tempHomework.save()
-                    _temp = 1
+                        tempHomework = _form.save(commit=False)
 
-                homework.save()
+                        tempHomework.title = form.cleaned_data.get("title")
+                        tempHomework.content = form.cleaned_data.get("content")
+                        tempHomework.end_date = form.cleaned_data.get("end_date")
+                        tempHomework.end_clock = form.cleaned_data.get("end_clock")
+                        tempHomework.homework_code = temp
+                        tempHomework.classroom_id = id
+                        tempHomework.submit_count = 0
+                        tempHomework.is_end = 0
+                        tempHomework.save()
+                        _temp = 1
 
-        return redirect("index")
+                    homework.save()
+
+            return redirect("index")
+
+        else:
+            form = NewHomeworkForm()
+
+        return render(request, "addhomework.html", {"classroom": classroom, "form": form})
 
     else:
-        form = NewHomeworkForm()
+        messages.info(request, "Bu sayfaya erişim hakkınız yok!")
+        return render(request, "about.html")
 
-    return render(request, "addhomework.html", {"classroom": classroom, "form": form})
 
-
+@login_required(login_url="user:login")
 def classDetail(request, id):
     # classroom = NewClass.objects.filter(id=id).first()
     classroom = get_object_or_404(NewClass, id=id)
+    print(request.user.id)
 
-    homework = Homework.objects.all()
+    if classroom.class_teacher_name_id == request.user.id:
 
-    post = ClassPost.objects.order_by('-created_date')
+        homework = Homework.objects.all()
 
-    form = ClassPostForm(request.POST or None, request.FILES or None)
+        post = ClassPost.objects.order_by('-created_date')
 
-    deneme = request.GET.get('kral')
+        form = ClassPostForm(request.POST or None, request.FILES or None)
+
+        deneme = request.GET.get('kral')
+
+        if form.is_valid():
+            post = form.save(commit=False)
+            post.author = request.user
+            post.classroom_id = id
+            post.save()
+
+            return redirect("/classes/{}".format(id))
+
+        homework2 = ClassHomework.objects.filter(classroom_id=id).values()
+
+        return render(request, "classDetail.html",
+                      {"classroom": classroom, "homework": homework, "id": id, "form": form, "post": post})
+
+    else:
+        messages.info(request, "Bu sayfaya erişiminiz yok!")
+        return render(request, "about.html")
 
 
-    if form.is_valid():
-        post = form.save(commit=False)
-        post.author = request.user
-        post.classroom_id = id
-        post.save()
-
-        return redirect("/classes/{}".format(id))
-
-    homework2 = ClassHomework.objects.filter(classroom_id=id).values()
-
-    return render(request, "classDetail.html",
-                  {"classroom": classroom, "homework": homework, "id": id, "form": form, "post": post})
-
-
+@login_required(login_url="user:login")
 def joinClass(request, keyword):
     student = NewContentForm(request.POST or None)
 
@@ -152,6 +169,7 @@ def joinClass(request, keyword):
         messages.info(request, "Düzgün bir sınıf kodu giriniz.")
 
 
+@login_required(login_url="user:login")
 def homeworks(request):
     class_content = ClassContent.objects.filter(student_name_id=request.user).values()
     _temp = ""
@@ -188,57 +206,82 @@ def homeworks(request):
                    "current_date": date.today(), "current_time": zaman})
 
 
+@login_required(login_url="user:login")
 def submitHomework(request, id):
+    now = datetime.now()
+
+    zaman = time(int(now.strftime(('%H'))), int(now.strftime(('%M'))), int(now.strftime(('%S'))))
     homework = get_object_or_404(ClassHomework, id=id)
-    allHomework = ClassHomework.objects.filter(student_id=request.user)
-    if request.method == "POST":
-        form = _NewHomeworkDeliveryForm(request.POST, request.FILES, instance=homework)
-        if form.is_valid():
-            file = form.save(commit=False)
-            file.student_id = request.user
-            file.is_deliver = 1
-            file.save()
-            messages.success(request, "Ödev başarıyla teslim edildi.")
-            return redirect('/classes/homeworks/')
+
+    if homework.student_name == request.user.username:
+
+        if homework.end_date > date.today():
+            if homework.end_clock > zaman:
+
+                allHomework = ClassHomework.objects.filter(student_id=request.user)
+                if request.method == "POST":
+                    form = _NewHomeworkDeliveryForm(request.POST, request.FILES, instance=homework)
+                    if form.is_valid():
+                        file = form.save(commit=False)
+                        file.student_id = request.user
+                        file.is_deliver = 1
+                        file.save()
+                        messages.success(request, "Ödev başarıyla teslim edildi.")
+                        return redirect('/classes/homeworks/')
+                else:
+                    form = _NewHomeworkDeliveryForm()
+                form = _NewHomeworkDeliveryForm()
+                return render(request, 'submitHomework.html', {
+                    'form': form, 'homework': homework, 'allHomework': allHomework
+                })
+
+            else:
+                messages.warning(request, "Ödevin süresi geçmiş. Bu ödevi teslim edemezsiniz!")
+                return render(request, "about.html")
+        else:
+            messages.warning(request, "Ödevin süresi geçmiş. Bu ödevi teslim edemezsiniz!")
+            return render(request, "about.html")
+
     else:
-        form = _NewHomeworkDeliveryForm()
-    form = _NewHomeworkDeliveryForm()
-    return render(request, 'submitHomework.html', {
-        'form': form, 'homework': homework, 'allHomework': allHomework
-    })
+        messages.info(request, "Bu ödeve erişim hakkınız yok!")
+        return render(request, "about.html")
 
 
+@login_required(login_url="user:login")
 def homeworkDetail(request, slug):
     sa = get_object_or_404(Homework, homework_code=slug)
     homeworkss = ClassHomework.objects.filter(homework_code=sa.homework_code)
-    """form2 = NoteForm(request.POST or None, instance=sa)
-    print("asdsad", form2)
-    form = NoteForm(request.POST or None)
-    if form.is_valid():
-        print("burda")
-        print(form)
-        username = form.cleaned_data.get("note")
-        print(username)"""
+
+    if request.GET.get("deneme41"):
+        print(request.GET.get("sa"))
 
     return render(request, 'homeworkDetail.html', {'homeworkss': homeworkss, 'sa': sa})
 
 
+@login_required(login_url="user:login")
 def dashboard(request, id):
     classroom = get_object_or_404(NewClass, id=id)
 
-    code = classroom.class_code
+    if classroom.class_teacher_name_id == request.user.id:
 
-    class_content = ClassContent.objects.filter(classroom_id=id)
+        code = classroom.class_code
 
-    homeworks = Homework.objects.filter(classroom_id=id)
+        class_content = ClassContent.objects.filter(classroom_id=id)
 
-    posts = ClassPost.objects.filter(classroom_id=id)
+        homeworks = Homework.objects.filter(classroom_id=id)
 
-    return render(request, "classDashboard.html",
-                  {"classroom": classroom, "class_content": class_content, "posts": posts, "code": code,
-                   "homeworks": homeworks})
+        posts = ClassPost.objects.filter(classroom_id=id)
+
+        return render(request, "classDashboard.html",
+                      {"classroom": classroom, "class_content": class_content, "posts": posts, "code": code,
+                       "homeworks": homeworks})
+
+    else:
+        messages.info(request, "Bu sayfaya erişim izniniz yok!")
+        return render(request, "about.html")
 
 
+@login_required(login_url="user:login")
 def addPost(request):
     form = ClassPostForm(request.POST or None, request.FILES or None)
 
@@ -251,33 +294,46 @@ def addPost(request):
         return redirect("index")
 
 
+@login_required(login_url="user:login")
 def updatePost(request, id):
     post = get_object_or_404(ClassPost, id=id)
-    form = ClassPostForm(request.POST or None, request.FILES or None, instance=post)
+    if post.author_id == request.user.id:
+        form = ClassPostForm(request.POST or None, request.FILES or None, instance=post)
 
-    if form.is_valid():
-        article = form.save(commit=False)
+        if form.is_valid():
+            article = form.save(commit=False)
 
-        article.author = request.user
-        article.save()
+            article.author = request.user
+            article.save()
 
-        messages.success(request, "Makale Başarıyla Güncellendi.")
-        return redirect("/classes/{}/dashboard".format(post.classroom_id))
+            messages.success(request, "Makale Başarıyla Güncellendi.")
+            return redirect("/classes/{}/dashboard".format(post.classroom_id))
 
-    return render(request, "update_post.html", {"form": form})
+        return render(request, "update_post.html", {"form": form})
+
+    else:
+        messages.info(request, "Bu sayfaya erişim hakkınız yok!")
+        return render(request, "about.html")
 
 
 def updateHomework(request, slug):
     pass
 
 
+@login_required(login_url="user:login")
 def deletePost(request, id):
     post = get_object_or_404(ClassPost, id=id)
-    post.delete()
-    messages.success(request, "Gönderi Başarıyla Silindi.")
-    return redirect("/classes/{}/dashboard".format(post.classroom_id))
+    if post.author_id == request.user.id:
+
+        post.delete()
+        messages.success(request, "Gönderi Başarıyla Silindi.")
+        return redirect("/classes/{}/dashboard".format(post.classroom_id))
+    else:
+        messages.info(request, "Bu sayfaya erişim hakkınız yok!")
+        return render(request, "about.html")
 
 
+@login_required(login_url="user:login")
 def deleteHomework(request, slug):
     classroom = get_object_or_404(Homework, homework_code=slug)
     ClassHomework.objects.filter(homework_code=slug).delete()
